discord-bot/requirements.txt
nextcord>=2.6.0
jellyfin-apiclient-python>=1.9.2
python-dotenv>=1.0.0
PyNaCl>=1.5.0